requirements.txt
fastapi
uvicorn
langgraph
langchain
groq
paddleocr
chromadb
pandas
numpy
python-dotenv
pillow
